+++ HISTORY.rst
@@ -56,7 +56,8 @@
 - Add new revertsettings management command.
 - Add new permission to edit setting via the UI.
 - Renamed setting LOCK_MANAGER_DEFAULT_BACKEND to LOCK_MANAGER_BACKEND.
-- Add help text to settings.
+- Add help texts to more setting options.
+- Add ACL support for metadata types.
 
 3.0.1 (2018-07-08)
 =================